please revert commit aaa4e02c9baea24371ab021de4f00d69cb681095
that patch was a quick workaround for alpha, the issue beeing fixed by
commit 1181b5904272b884afda429461765142e72e34fc

+++ include/sys/vfs.h
@@ -32,17 +32,17 @@ struct statfs {
 #else /* _BITSIZE == 64 */
 
 struct statfs {
-        uint32_t f_type;
-        uint32_t f_bsize;
+        uint64_t f_type;
+        uint64_t f_bsize;
         uint64_t f_blocks;
         uint64_t f_bfree;
         uint64_t f_bavail;
         uint64_t f_files;
         uint64_t f_ffree;
         __kernel_fsid_t f_fsid;
-        uint32_t f_namelen;
-        uint32_t f_frsize;
-        uint32_t f_spare[5];
+        uint64_t f_namelen;
+        uint64_t f_frsize;
+        uint64_t f_spare[5];
 };
 
 #endif /* _BITSIZE */
